Move test helpers into test files
